Separated code in headers and minor changes
Removed useless functions separated all in headers
