+++ HuffmanBinaryTree/LLBasicFunctions.cpp
@@ -56,37 +56,7 @@ LL* addToEnd(LL** head, int x)
 		return (*head);
 	}
 }
-LL* addToStart(int x)
-{
-	LL* head = new LL;
-	head->x = x;
-	head->next = NULL;
-	return head;
-}
-void insertNodeinSorted(LL** head, int x)
-{
-	LL* temp = new LL;
-	temp->x = x;
-	if (x > (*head)->x)
-	{
-		temp->next = (*head);
-		(*head) = temp;
-	}
-	else {
-		LL* t = (*head);
-		while (t->next->x > x && t->next->next->x <= x)t = t->next;
-		if (t->next == NULL) {
-			temp->next = NULL;
-			t->next = temp;
-		}
-		else {
-			temp->x = x;
-			temp->next = t->next;
-			t->next = temp;
-		}
-	}
-	
-}
+
 void removeLastElement(LL** head)
 {
 	LL* temp = NULL;
@@ -108,37 +78,7 @@ int peek(LL* head)
 		t = t->next;
 	return t->x;
 }
-double arithmeticMean(LL* head)
-{
-	double as = 0;
-	int counter(0);
 
-	LL* t = head;
-	while (t != NULL)
-	{
-		counter++;
-		as += t->x;
-		t = t->next;
-	}
-	return as /= counter;
-}
-void insert(LL* head, int index, int x)
-{
-	LL* t = head;
-	LL* temp = NULL;
-	int counter(0);
-	while (true)
-	{
-		counter++;
-		if (counter == index + 1)break;
-		t = t->next;
-	}
-
-	LL* novi = new LL;
-	novi->x = x;
-	novi->next = t->next;
-	t->next = novi;
-}
 
 
 void insertionSortDesc(struct LL** head)
@@ -150,20 +90,6 @@ void insertionSortDesc(struct LL** head)
 	{
 		struct LL* next = current->next;
 		sortedInsertDesc(&sorted, current);
-		current = next;
-	}
-
-	*head = sorted;
-}
-void insertionSortAsc(struct LL** head)
-{
-	struct LL* sorted = NULL;
-
-	struct LL* current = *head;
-	while (current != NULL)
-	{
-		struct LL* next = current->next;
-		sortedInsertAsc(&sorted, current);
 		current = next;
 	}
 
@@ -189,27 +115,6 @@ void sortedInsertDesc(struct LL** head, struct LL* new_node)
 		current->next = new_node;
 	}
 }
-void sortedInsertAsc(struct LL** head, struct LL* new_node)
-{
-	struct LL* current;
-	if (*head == NULL || (*head)->x >= new_node->x)
-	{
-		new_node->next = *head;
-		*head = new_node;
-	}
-	else
-	{
-		current = *head;
-		while (current->next != NULL &&
-			current->next->x < new_node->x)
-		{
-			current = current->next;
-		}
-		new_node->next = current->next;
-		current->next = new_node;
-	}
-}
-
 
 void fillListFromArray(LL* head, std::array<int, 26>arr,double maxProbability)
 {
